requirements.txt
spotipy==2.24.0
pandas==2.2.2
pyarrow==16.1.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
apscheduler==3.10.4
python-dotenv==1.0.1
schedule==1.2.2